chore: move isFucusable
