+++ app/src/hooks/useTempoEditorKeyboardShortcut.tsx
@@ -7,7 +7,7 @@ import {
   useResetTempoSelection,
 } from "../actions/tempo"
 import { TempoEventsClipboardDataSchema } from "../clipboard/clipboardTypes"
-import { isFocusable } from "../components/KeyboardShortcut/isFocusable"
+import { isFocusable } from "../helpers/isFocusable"
 import { readClipboardData } from "../services/Clipboard"
 import { useKeyboardShortcut } from "./useKeyboardShortcut"
 import { useTempoEditor } from "./useTempoEditor"
